게시물 작성 페이지 css 수정
