+++ src/routes/layouts/CreatPostLayout.tsx
@@ -1,0 +1,11 @@
+import { Outlet } from 'react-router';
+import Header from '../../components/common/Header';
+
+export default function CreatePostLayout() {
+  return (
+    <>
+      <Header />
+      <Outlet />
+    </>
+  );
+}
